Add NowWidget feature with localized strings and session management
- Introduced NowWidget to display event phases: before, during, and after events.
- Implemented localized strings for better user experience in multiple languages.
- Created NowEntry structure to manage timeline entries for widget updates.
- Added session handling logic to fetch and display current and upcoming sessions.
- Enhanced UI with SwiftUI for a more interactive widget experience.
