+++ iPlayground/NowWidget/Assets.xcassets/WidgetBackground.colorset/Contents.json
@@ -1,6 +1,15 @@
 {
   "colors" : [
     {
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "alpha" : "1.000",
+          "blue" : "0xFF",
+          "green" : "0x7B",
+          "red" : "0x4F"
+        }
+      },
       "idiom" : "universal"
     }
   ],
