+++ iPlayground/NowWidget/Assets.xcassets/iPlaygroundBlue.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0xFF",
-          "green" : "0x7B",
-          "red" : "0x4F"
+          "blue" : "0xDF",
+          "green" : "0x47",
+          "red" : "0x3B"
         }
       },
       "idiom" : "universal"
@@ -24,8 +24,8 @@
         "components" : {
           "alpha" : "1.000",
           "blue" : "0xFF",
-          "green" : "0x7B",
-          "red" : "0x4F"
+          "green" : "0x85",
+          "red" : "0x7B"
         }
       },
       "idiom" : "universal"
