+++ iPlayground/NowWidget/Assets.xcassets/iPlaygroundBlueBackground.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0xFF",
-          "green" : "0x7B",
-          "red" : "0x4F"
+          "blue" : "0xEF",
+          "green" : "0x57",
+          "red" : "0x4B"
         }
       },
       "idiom" : "universal"
@@ -23,9 +23,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0xFF",
-          "green" : "0x7B",
-          "red" : "0x4F"
+          "blue" : "0x70",
+          "green" : "0x36",
+          "red" : "0x2D"
         }
       },
       "idiom" : "universal"
