+++ iPlayground/NowWidget/Assets.xcassets/iPlaygroundPink.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0xFF",
-          "green" : "0x7B",
-          "red" : "0x4F"
+          "blue" : "0x9A",
+          "green" : "0x45",
+          "red" : "0xD4"
         }
       },
       "idiom" : "universal"
@@ -23,9 +23,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0xFF",
-          "green" : "0x7B",
-          "red" : "0x4F"
+          "blue" : "0xBD",
+          "green" : "0x79",
+          "red" : "0xFF"
         }
       },
       "idiom" : "universal"
