+++ iPlayground/NowWidget/Assets.xcassets/iPlaygroundPinkBackground.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0xFF",
-          "green" : "0x7B",
-          "red" : "0x4F"
+          "blue" : "0xA7",
+          "green" : "0x5F",
+          "red" : "0xEE"
         }
       },
       "idiom" : "universal"
@@ -23,9 +23,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0xFF",
-          "green" : "0x7B",
-          "red" : "0x4F"
+          "blue" : "0x5E",
+          "green" : "0x3A",
+          "red" : "0x7A"
         }
       },
       "idiom" : "universal"
