+++ iPlayground/NowWidget/NowWidgetBundle.swift
@@ -5,12 +5,12 @@
 //  Created by ethanhuang on 2025/8/28.
 //
 
-import WidgetKit
 import SwiftUI
+import WidgetKit
 
 @main
 struct NowWidgetBundle: WidgetBundle {
-    var body: some Widget {
-        NowWidget()
-    }
+  var body: some Widget {
+    NowWidget()
+  }
 }
